Finished machine learning
Finished machine learning using gaussian mixture models. Fixed a bug when calculating maxDelta that caused it to always calculate 0. Fixed a bug when calculating the new mean of the second gaussian blob where the gamma variance sum was used instead of the gamma product sum. Added getters for all sums and weights to BimodalModel for debugging purposes. Added these new functions to the keywords.txt file to enable text highlighting on the arduino. Removed ITERS after realizing that the model was accidently being trained to find the WORST fit instead of the BEST fit... oops. ITERS was meant to average 20 models to get the divider, but after changing it to focus on the best fit only on model was needed. Removed threshold for random values because they didn't need it. Changed DELTA_THRESH to 0.05 because numbers below this caused very long learning times and the blobs seemed to converge at around this value of delta. Removed unneeded functions and variables from the program.